Split generic utils/utils.py into purpose-named modules
Replace the catch-all utils/utils.py with:
- utils/image_utils.py  (coord_reshape, coord2pixel_reverse, resolution_alignment, l2_norm)
- utils/brand_utils.py  (brand_converter)
utils/web_utils.py is unchanged. Update importers in modules/ and the README.
