+++ modules/crp_classifier.py
@@ -8,7 +8,7 @@ from lxml import html
 import io
 import os
 import numpy as np
-from utils.utils import coord2pixel_reverse
+from utils.image_utils import coord2pixel_reverse
 from networks.crp_models import KNOWN_MODELS
 import logging
