+++ modules/logo_matching.py
@@ -1,6 +1,7 @@
 from PIL import Image, ImageOps
 from torchvision import transforms
-from utils.utils import brand_converter, resolution_alignment, l2_norm
+from utils.brand_utils import brand_converter
+from utils.image_utils import resolution_alignment, l2_norm
 from networks.siamese_models import KNOWN_MODELS
 from networks.ocr.models.model_builder import ModelBuilder
 from networks.ocr.utils.labelmaps import get_vocabulary
